Updated gp_sampler testing. Next task is to make sure that this works with travis.
