+++ demo/demo_gp_python_api.py
@@ -23,7 +23,6 @@ from example_processes import simulate_measurement
 import matplotlib.pyplot as pl
 
 
-# TEST TRAVIS
 def main():
 
     # Set up a sampling problem
