+++ setup.cfg
@@ -1,2 +1,3 @@
 [pytest]
 addopts = --doctest-modules --ignore=setup.py
+norecursedirs = build docs demos .tox .git
